[#1477] added some resource level functions to check folder/file operations
